Move SQL_Stuff class to DataControlsLib project and add using statement where needed
